order desing is completed, also added userId to signuture but we already have username. thats why we will username for the user identity with username.

+++ src/main/java/com/onlineShop/bootcamp/dto/order/OrderItemResponse.java
@@ -16,7 +16,7 @@ public class OrderItemResponse {
     private Long productId;
     private String productName;
     private Integer quantity;
-    private BigDecimal perItemPrice;
+    private Double perItemPrice;
     private Double subTotal;
 
 }
