+++ src/main/java/com/onlineShop/bootcamp/dto/product/ProductRequest.java
@@ -16,7 +16,7 @@ public class ProductRequest {
     private Long supplierId;
     private String typeName;
     private Integer amountGrams;
-    private BigDecimal priceGbp;
+    private Double priceGbp;
     private BrewColor brewColor;
     private Integer caffeineMgPerG;
     private String mainMedicinalUse;
