+++ src/main/java/com/onlineShop/bootcamp/dto/product/ProductResponse.java
@@ -17,7 +17,7 @@ public class ProductResponse {
     private Long supplierId;
     private String typeName;
     private Integer amountGrams;
-    private BigDecimal priceGbp;
+    private Double priceGbp;
     private BrewColor brewColor;
     private Integer caffeineMgPerG;
     private String mainMedicinalUse;
